--- frontend/src/components/Header.tsx
@@ -1,5 +1,0 @@
-function Header() {
-  return <h1>prompt injection demo</h1>;
-}
-
-export default Header;
